+++ common/mr/ReduceAverageWithThreshold.java
@@ -2,11 +2,11 @@ package common.mr;
 
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.mapreduce.Reducer;
-import org.apache.hadoop.io.DoubleWritable;
+import org.apache.hadoop.io.FloatWritable;
 import org.apache.hadoop.io.Text;
 import java.io.IOException;
 
-public class ReduceAverageWithThreshold extends Reducer<Text, DoubleWritable, Text, DoubleWritable> {
+public class ReduceAverageWithThreshold extends Reducer<Text, FloatWritable, Text, FloatWritable> {
     Integer M;
 
     @Override
@@ -16,16 +16,16 @@ public class ReduceAverageWithThreshold extends Reducer<Text, DoubleWritable, Te
     }
 
     @Override
-    public void reduce(Text key, Iterable<DoubleWritable> values, Context context) throws IOException, InterruptedException {
-        double sum = 0.0;
+    public void reduce(Text key, Iterable<FloatWritable> values, Context context) throws IOException, InterruptedException {
+        Float sum = 0.0;
         long cnt = 0;
-        for (DoubleWritable val : values) {
+        for (FloatWritable val : values) {
             sum += val.get();
             cnt += 1;
         }
         // output only if items count is over the threshold
         if (cnt >= this.M) {
-            context.write(key, new DoubleWritable(sum / cnt));
+            context.write(key, new FloatWritable(sum / cnt));
         }
     }
 }
